Change: Works with type int

+++ Task01/Program.cs
@@ -1,8 +1,8 @@
 ﻿// Задача 1. Вывести квдрат числа
 
-double Square (double x)
+int Square (int x)
 {
-    double result =  x*x;
+    int result =  x*x;
     return result;
 }
 
@@ -10,6 +10,9 @@ Console.WriteLine("Введите число");
 string Number = Console.ReadLine();
 //int numb = Convert.ToInt32(Number);
 
-double numb = Square(Convert.ToDouble(Number)); 
+
+int numb = Square(Convert.ToInt32(Number)); 
 
 Console.WriteLine($"Квадрат введенного числа равен: {numb}");
+
+
